Flatten ComputeFirewallPolicyRule directory
