--- pkg/controller/direct/compute/firewallpolicyrule/mapper.go
@@ -1,94 +1,0 @@
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package firewallpolicyrule
-
-import (
-	pb "cloud.google.com/go/compute/apiv1/computepb"
-	krm "github.com/GoogleCloudPlatform/k8s-config-connector/apis/compute/v1beta1"
-	refs "github.com/GoogleCloudPlatform/k8s-config-connector/apis/refs/v1beta1"
-
-	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct"
-)
-
-func ComputeFirewallPolicyRuleSpec_TargetResources_ToProto(mapCtx *direct.MapContext, in []*refs.ComputeNetworkRef) []string {
-	if in == nil {
-		return nil
-	}
-	var out []string
-	for _, i := range in {
-		if i == nil {
-			continue
-		}
-		if i.External == "" {
-			mapCtx.Errorf("reference %s was not pre-resolved", i.Name)
-		}
-		out = append(out, i.External)
-	}
-	return out
-}
-
-func ComputeFirewallPolicyRuleSpec_TargetServiceAccounts_ToProto(mapCtx *direct.MapContext, in []*refs.IAMServiceAccountRef) []string {
-	if in == nil {
-		return nil
-	}
-	var out []string
-	for _, i := range in {
-		if i == nil {
-			continue
-		}
-		if i.External == "" {
-			mapCtx.Errorf("reference %s was not pre-resolved", i.Name)
-		}
-		out = append(out, i.External)
-	}
-	return out
-}
-
-func ComputeFirewallPolicyRuleSpec_TargetResources_FromProto(mapCtx *direct.MapContext, in []string) []*refs.ComputeNetworkRef {
-	if in == nil {
-		return nil
-	}
-	var out []*refs.ComputeNetworkRef
-	for _, i := range in {
-		out = append(out, &refs.ComputeNetworkRef{
-			External: i,
-		})
-	}
-	return out
-}
-
-func ComputeFirewallPolicyRuleSpec_TargetServiceAccounts_FromProto(mapCtx *direct.MapContext, in []string) []*refs.IAMServiceAccountRef {
-	if in == nil {
-		return nil
-	}
-	var out []*refs.IAMServiceAccountRef
-	for _, i := range in {
-		out = append(out, &refs.IAMServiceAccountRef{
-			External: i,
-		})
-	}
-	return out
-}
-
-func ComputeFirewallPolicyRuleStatus_ToProto(mapCtx *direct.MapContext, in *krm.ComputeFirewallPolicyRuleStatus) *pb.FirewallPolicyRule {
-	if in == nil {
-		return nil
-	}
-	out := &pb.FirewallPolicyRule{}
-	out.Kind = in.Kind
-	out.RuleTupleCount = direct.PtrInt64ToPtrInt32(in.RuleTupleCount)
-	// MISSING: TargetSecureTags
-	return out
-}
